refactor(events): move pure title helpers into src/lib/eventTitle
The listing-quality checks need composeEventTitle / EVENT_TITLE_DEFAULTS /
addressPlaceName to recognise an auto-filled title, and they live in
src/lib/eventQuality. src/lib/ imports nothing from src/collections/ (rule 4
in .claude/rules/project-structure.md), so the pure half of the title hook
moves to src/lib/eventTitle/compose.ts. The request-scoped half —
resolveTitleTemplates + eventTitleBeforeChange — stays with the collection.

Refs #609
